fixed issue where deleting images could segfault
In some cases, deleting an image would cause the next iteration of the
loop in xfinishdraw to attempt to access the next image, using the
previously freed pointer as a base. This patch causes deleting an image
to rest the im pointer to term.images each time an image is deleted.