fixed some odd behavior
commenting out sixel_parser_deinit() in the DCS handler seemed to cause
issues with correctly making enough vertical space for images.

The issue where the line on which a command was entered would disappear
has not been fixed, but it only occurs with lsix, which makes me think
that it may be a control character being emitted from that. Using
sixelconv or cat-ing a sixel does not re-produce the problem.