fixed bug in handling of S and T CSI escapes
When a S or T CSI escape was encountered, the lines which were scrolled
away would be deleted from the scrollback buffer. This has been
corrected - the lines are now preseved.

This fixes a bug where issuing `clear` followed by `lsix` would cause
the line on which the `lsix` was issued to disappear from the scrollback
buffer.

Note that the line may scroll out of view and thus dissapear, but it
will now be preserved in the scrollback buffer.